ci: provide composite action and check it using itself (#3)
* ci: provide composite action and check it using itself

By adding `action.yml` to the top level, this repo can be used as an action,
without having to know that is uses go under the hood.

Of course we eat the dog food: added a job to the github workflow that uses itself as an action to lint the action.yml file.

* docs: add usage example and improve readme

* docs: add workflow comment